tests/wpt/web-platform-tests/html/semantics/popovers/popover-events.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popover events</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popover-utils.js"></script>

<div popover>Popover</div>

<script>
window.onload = () => {
  for(const method of ["listener","attribute"]) {
    promise_test(async t => {
      const popover = document.querySelector('[popover]');
      assert_false(popover.matches(':open'));
      let showCount = 0;
      let hideCount = 0;
      function showListener(e) {
        assert_true(e.target.matches(':closed'),'The popover should be in the :closed state when the popovershow event fires.');
        assert_false(e.target.matches(':open'),'The popover should *not* be in the :open state when the popovershow event fires.');
        ++showCount;
      };
      function hideListener(e) {
        assert_true(e.target.matches(':open'),'The popover should be in the :open state when the popoverhide event fires.');
        assert_false(e.target.matches(':closed'),'The popover should *not* be in the :closed state when the popoverhide event fires.');
        ++hideCount;
      };
      switch (method) {
        case "listener":
          const controller = new AbortController();
          const signal = controller.signal;
          t.add_cleanup(() => controller.abort());
          document.addEventListener('popovershow',showListener, {signal});
          document.addEventListener('popoverhide',hideListener, {signal});
          break;
        case "attribute":
          assert_false(popover.hasAttribute('onpopovershow'));
          assert_false(popover.hasAttribute('onpopoverhide'));
          t.add_cleanup(() => popover.removeAttribute('onpopovershow'));
          t.add_cleanup(() => popover.removeAttribute('onpopoverhide'));
          popover.onpopovershow = showListener;
          popover.onpopoverhide = hideListener;
          break;
        default: assert_unreached();
      }
      assert_equals(0,showCount);
      assert_equals(0,hideCount);
      popover.showPopover();
      assert_true(popover.matches(':open'));
      assert_equals(1,showCount);
      assert_equals(0,hideCount);
      await waitForRender();
      assert_true(popover.matches(':open'));
      popover.hidePopover();
      assert_false(popover.matches(':open'));
      assert_equals(1,showCount);
      assert_equals(1,hideCount);
      await waitForRender();
      // No additional events after animation frame
      assert_false(popover.matches(':open'));
      assert_equals(1,showCount);
      assert_equals(1,hideCount);
    }, `Popovershow and popoverhide events (${method}) get properly dispatched for popovers`);
  }

  promise_test(async t => {
    const popover = document.querySelector('[popover]');
    const controller = new AbortController();
    const signal = controller.signal;
    t.add_cleanup(() => controller.abort());
    let cancel = true;
    popover.addEventListener('popovershow',(e) => {
      if (cancel)
        e.preventDefault();
    }, {signal});
    assert_false(popover.matches(':open'));
    popover.showPopover();
    assert_false(popover.matches(':open'),'The "popovershow" event should be cancelable');
    cancel = false;
    popover.showPopover();
    assert_true(popover.matches(':open'));
    popover.hidePopover();
    assert_false(popover.matches(':open'));
  }, 'Popovershow event is cancelable');
};
</script>
